some changes

+++ Domain/Entity/BorrowHistory.cs
@@ -2,9 +2,9 @@ namespace Domain.Entity;
 
 public class BorrowHistory
 {
-    public int Id { get; set; }
-    public int UserId { get; set; }
-    public int BookId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
+    public long BookId { get; set; }
     public DateTime BorrowDate { get; set; } = DateTime.UtcNow;
     public DateTime? ReturnedDate { get; set; }
     public DateTime DueDate { get; set; }
